Fix bug related to path dependent types.
Previously, the rewrite would insert names like `Value` in the position
of a path dependent type such as `Day.Value` even if the `Value` in
scope had a conflicting type `Bool.Value`. Now, when querying the
scope we take into account the prefix from where the symbols in scope
origin.

The implementation in this PR is probably incorrect for some cases. I
had to use a hacky workaround for the custom `PrettyType` type. Either
way, this diff makes the test suite pass and allow us to move on.