Handle case where term name and type name resolve to different symbols.
Previously, the rewrite would insert a name `Map` as long as either the
term name or type name would resolve to a desired symbol. This resulted
in compile errors when the term name and type name resolved to
conflicting symbols. Now, we only insert the unqualified `Map` name as
long as the term and type name don't resolve to conflicting symbols.